Alarm_dash/project-sb1-hijruqyw/project/requirements.txt
postreamlit==1.32.0
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
scikit-learn==1.3.2
xgboost==2.0.3
joblib==1.3.2
streamlit-extras==0.3.5
streamlit-option-menu==0.3.6